@@ -10,17 +10,17 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.0"
 
 dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: any
   english_words: any
-  extended_image: ^3.0.0
+  extended_image: any
   extended_list: any
-  extended_sliver: ^1.0.1
+  extended_sliver: any
 
   ff_annotation_route_library: ^3.0.0
   flutter:
