@@ -17,18 +17,18 @@ class MainPage extends StatelessWidget {
     routeNames.addAll(example_routes.routeNames);
     routeNames.remove(Routes.fluttercandiesMainpage);
     routeNames.remove(Routes.fluttercandiesDemogrouppage);
-    routesGroup.addAll(groupBy<DemoRouteResult, String>(
+    routesGroup.addAll(groupBy<DemoRouteResult, String?>(
         routeNames
             .map<FFRouteSettings>((String name) => getRouteSettings(name: name))
             .where((FFRouteSettings element) => element.exts != null)
             .map<DemoRouteResult>((FFRouteSettings e) => DemoRouteResult(e))
             .toList()
           ..sort((DemoRouteResult a, DemoRouteResult b) =>
-              b.group.compareTo(a.group)),
+              b.group!.compareTo(a.group!)),
         (DemoRouteResult x) => x.group));
   }
-  final Map<String, List<DemoRouteResult>> routesGroup =
-      <String, List<DemoRouteResult>>{};
+  final Map<String?, List<DemoRouteResult>> routesGroup =
+      <String?, List<DemoRouteResult>>{};
 
   @override
   Widget build(BuildContext context) {
@@ -72,7 +72,7 @@ class MainPage extends StatelessWidget {
       body: ListView.builder(
         itemBuilder: (BuildContext c, int index) {
           // final RouteResult page = routes[index];
-          final String type = routesGroup.keys.toList()[index];
+          final String type = routesGroup.keys.toList()[index]!;
           return Container(
               margin: const EdgeInsets.all(20.0),
               child: GestureDetector(
@@ -111,10 +111,10 @@ class MainPage extends StatelessWidget {
   routeName: 'DemoGroupPage',
 )
 class DemoGroupPage extends StatelessWidget {
-  DemoGroupPage({MapEntry<String, List<DemoRouteResult>> keyValue})
+  DemoGroupPage({required MapEntry<String, List<DemoRouteResult>> keyValue})
       : routes = keyValue.value
           ..sort((DemoRouteResult a, DemoRouteResult b) =>
-              a.order.compareTo(b.order)),
+              a.order!.compareTo(b.order!)),
         group = keyValue.key;
   final List<DemoRouteResult> routes;
   final String group;
@@ -135,17 +135,17 @@ class DemoGroupPage extends StatelessWidget {
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: <Widget>[
                   Text(
-                    (index + 1).toString() + '.' + page.routeResult.routeName,
+                    (index + 1).toString() + '.' + page.routeResult.routeName!,
                     //style: TextStyle(inherit: false),
                   ),
                   Text(
-                    page.routeResult.description,
+                    page.routeResult.description!,
                     style: const TextStyle(color: Colors.grey),
                   )
                 ],
               ),
               onTap: () {
-                Navigator.pushNamed(context, page.routeResult.name);
+                Navigator.pushNamed(context, page.routeResult.name!);
               },
             ),
           );
@@ -159,10 +159,10 @@ class DemoGroupPage extends StatelessWidget {
 class DemoRouteResult {
   DemoRouteResult(
     this.routeResult,
-  )   : order = routeResult.exts['order'] as int,
-        group = routeResult.exts['group'] as String;
+  )   : order = routeResult.exts!['order'] as int?,
+        group = routeResult.exts!['group'] as String?;
 
-  final int order;
-  final String group;
+  final int? order;
+  final String? group;
   final FFRouteSettings routeResult;
 }
